+++ clients/generator.c
@@ -49,7 +49,7 @@ pairs_st *pairs_generate(uint64_t number_of, size_t value_length)
   unsigned int x;
   pairs_st *pairs;
 
-  pairs= (pairs_st*)calloc(number_of + 1, sizeof(pairs_st));
+  pairs= (pairs_st*)calloc((size_t)number_of + 1, sizeof(pairs_st));
 
   if (!pairs)
     goto error;
